+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.8.23-exp.2] - 2022-04-21
+- Added support for parametric CodeBasedGraphicsTests.
+- Added support for Playmode CodeBasedGraphicsTests.
+- Made Graphics Test Results window TreeView scrollable
+- Added compatibility between CodeBasedGraphicsTests and the ReferenceImagesBase folder
+
 ## [7.8.23-exp.1] - 2022-04-11
 - Added support for new PS5 GraphicsDeviceType.